pyproject.toml
[project]
name = "context-engine"
version = "1.0.0"
description = "Code indexing and query engine for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Context Engine Team" }
]

dependencies = [
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-typescript>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-go>=0.23.0",
    "tree-sitter-rust>=0.23.0",
    "tree-sitter-java>=0.23.0",
    "fastmcp>=2.0.0",
    "watchdog>=4.0.0",
    "pathspec>=0.12.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
ce = "cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["engine"]
sources = ["cli.py", "server.py", "config.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
